+++ src/app/folder/play/constants.ts
@@ -1,7 +1,3 @@
-export const fieldNames = {
-	FOLDER_ID: 'folder_id',
-};
-
 export const MAX_COUNT_WINS = 20;
 
 export const noDataProps = {
